requirements-dev.txt
# Include base requirements
-r requirements.txt

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Type checking
mypy>=1.0.0
pandas-stubs>=2.0.0

# Linting and formatting
ruff>=0.1.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=2.0.0
